requirements.txt
kafka-python-ng==2.2.3
pandas==2.2.2
plotly==5.24.0
psycopg2-binary==2.9.9
pyspark==3.5.1
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.44.1
